Move error code map to internal package
