+++ internal/grpcerrorcodes/codes.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package protobuf
+package grpcerrorcodes
 
 import (
 	"go.uber.org/yarpc/yarpcerrors"
@@ -26,8 +26,8 @@ import (
 )
 
 var (
-	// _codeToGRPCCode maps all Codes to their corresponding gRPC Code.
-	_codeToGRPCCode = map[yarpcerrors.Code]codes.Code{
+	// YARPCCodeToGRPCCode maps all Codes to their corresponding gRPC Code.
+	YARPCCodeToGRPCCode = map[yarpcerrors.Code]codes.Code{
 		yarpcerrors.CodeOK:                 codes.OK,
 		yarpcerrors.CodeCancelled:          codes.Canceled,
 		yarpcerrors.CodeUnknown:            codes.Unknown,
@@ -45,5 +45,26 @@ var (
 		yarpcerrors.CodeUnavailable:        codes.Unavailable,
 		yarpcerrors.CodeDataLoss:           codes.DataLoss,
 		yarpcerrors.CodeUnauthenticated:    codes.Unauthenticated,
+	}
+
+	// GRPCCodeToYARPCCode maps all gRPC Codes to their corresponding Code.
+	GRPCCodeToYARPCCode = map[codes.Code]yarpcerrors.Code{
+		codes.OK:                 yarpcerrors.CodeOK,
+		codes.Canceled:           yarpcerrors.CodeCancelled,
+		codes.Unknown:            yarpcerrors.CodeUnknown,
+		codes.InvalidArgument:    yarpcerrors.CodeInvalidArgument,
+		codes.DeadlineExceeded:   yarpcerrors.CodeDeadlineExceeded,
+		codes.NotFound:           yarpcerrors.CodeNotFound,
+		codes.AlreadyExists:      yarpcerrors.CodeAlreadyExists,
+		codes.PermissionDenied:   yarpcerrors.CodePermissionDenied,
+		codes.ResourceExhausted:  yarpcerrors.CodeResourceExhausted,
+		codes.FailedPrecondition: yarpcerrors.CodeFailedPrecondition,
+		codes.Aborted:            yarpcerrors.CodeAborted,
+		codes.OutOfRange:         yarpcerrors.CodeOutOfRange,
+		codes.Unimplemented:      yarpcerrors.CodeUnimplemented,
+		codes.Internal:           yarpcerrors.CodeInternal,
+		codes.Unavailable:        yarpcerrors.CodeUnavailable,
+		codes.DataLoss:           yarpcerrors.CodeDataLoss,
+		codes.Unauthenticated:    yarpcerrors.CodeUnauthenticated,
 	}
 )
