+++ transport/grpc/handler.go
@@ -28,6 +28,7 @@ import (
 	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/api/transport"
 	"go.uber.org/yarpc/internal/bufferpool"
+	"go.uber.org/yarpc/internal/grpcerrorcodes"
 	"go.uber.org/yarpc/yarpcerrors"
 	"go.uber.org/zap"
 	"golang.org/x/net/context"
@@ -270,8 +271,8 @@ func handlerErrorToGRPCError(err error, responseWriter *responseWriter) error {
 		return unmarshalError(body)
 	}
 
-	grpcCode, ok := _codeToGRPCCode[yarpcStatus.Code()]
-	// should only happen if _codeToGRPCCode does not cover all codes
+	grpcCode, ok := grpcerrorcodes.YARPCCodeToGRPCCode[yarpcStatus.Code()]
+	// should only happen if grpcerrorcodes.YARPCCodeToGRPCCode does not cover all codes
 	if !ok {
 		grpcCode = codes.Unknown
 	}
