+++ transport/grpc/outbound.go
@@ -32,6 +32,7 @@ import (
 	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/api/peer"
 	"go.uber.org/yarpc/api/transport"
+	"go.uber.org/yarpc/internal/grpcerrorcodes"
 	intyarpcerrors "go.uber.org/yarpc/internal/yarpcerrors"
 	peerchooser "go.uber.org/yarpc/peer"
 	"go.uber.org/yarpc/peer/hostport"
@@ -211,7 +212,7 @@ func invokeErrorToYARPCError(err error, responseMD metadata.MD) error {
 	if !ok {
 		return yarpcerrors.FromError(err)
 	}
-	code, ok := _grpcCodeToCode[status.Code()]
+	code, ok := grpcerrorcodes.GRPCCodeToYARPCCode[status.Code()]
 	if !ok {
 		code = yarpcerrors.CodeUnknown
 	}
